requirements.txt
aiofiles==24.1.0
aiogram==3.16.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
code128==0.3
colorama==0.4.6
Django==5.1.4
frozenlist==1.5.0
idna==3.10
image==1.5.33
magic-filter==1.0.12
multidict==6.1.0
numpy==2.2.1
opencv-python==4.10.0.84
pillow==11.0.0
propcache==0.2.1
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
pylibdmtx==0.1.10
python-barcode==0.15.1
python-dotenv==1.0.1
qrcode==8.0
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
yarl==1.18.3